+++ src/classes/Map/index.ts
@@ -1,7 +1,7 @@
 import { DataType, IDataType } from '../DataType'
 import { EXCEEDS_MAX_CHILDREN } from '../../types/errors'
 import { PERMISSIONS, DEFAULT_PERMISSIONS } from '../../types/permissions'
-import { IDataTypeKind } from '..'
+import { IDataTypeKind } from '../DataType'
 
 export interface IMap extends IDataType {
   set(key: string, value: DataType): void
