+++ src/classes/Set/index.ts
@@ -2,7 +2,7 @@ import { List, IList } from '../List'
 import { PERMISSIONS, DEFAULT_PERMISSIONS } from '../../types/permissions'
 import { DataType } from '../DataType'
 import { CHILD_ALREADY_EXISTS } from '../../types/errors'
-import { IDataTypeKind } from '..'
+import { IDataTypeKind } from '../DataType'
 
 export interface ISet extends IList {
   add(child: DataType): DataType
