+++ src/classes/index.ts
@@ -5,13 +5,3 @@ export { Leaf, ILeaf } from './Leaf'
 export { List, IList } from './List'
 export { Map, IMap } from './Map'
 export { Set, ISet } from './Set'
-
-export enum IDataTypeKind {
-  Action = 'Action',
-  Choice = 'Choice',
-  DataType = 'DataType',
-  Leaf = 'Leaf',
-  List = 'List',
-  Map = 'Map',
-  Set = 'Set'
-}
